requirements_doc.txt
Sphinx>=1.4.1
sphinx-rtd-theme>=0.1.9
numpydoc>=0.8.0
